+++ R/onLoad.R
@@ -5,5 +5,6 @@ censobr_env <- new.env(parent = emptyenv()) # nocov start
 
   # data release
   censobr_env$data_release <- 'v0.6.0'
+  censobr_env$label_config_cache <- new.env(parent = emptyenv())
 
 } # nocov end